a tipsy biased halo finding implementation.  I couldn't figure out how to determine what type of file the simulation object came from.
But, for tipsy, pynbody looks for an AHF particle file or amiga.grp file.
Upon finding neither, pynbody searches your executable PATH environment
variable for 'AHFstep'.  If it finds that, it runs it for you.  This doesn't
take that long (3 minutes for 10 million particles?).  Then, it has to
load the particle file.  And I'm not proud of how slow the particle
file loading is, but it is not due to slow file read, it is due to
having to sort out a complicated file format.  I tried readlines
and still got caught up in sorting stuff out.

The units of the quantities in the loaded properties will be quite an
adventure to sort out.  You should be able to make a mass function now.