made satellite luminosity function plot
fixed schmidtlaw so that it works, but there is a python issue that
I don't understand described in Issue32.  This means that you can't
center halos using hybrid centering because something breaks when there's
multiple layers of SimArray-ness in my python2.6.  Something to do with
what the sphere filter returns for a Halo.