Refactor GadgetHDFSnap and related classes, and add ability for flexible naming so that reader works eg. for Gizmo files
This is a major refactorization to improve maintability and remove code duplication from GadgetHDFSnap and SubfindHDFSnap. It was triggered by trying to add the flexibility to load Gizmo files with the same reader (which is now possible). It involves breaking a lot of very long functions into smaller named functions to improve clarity, and removes a lot of duplicated code by using these new smaller functions.