Remove over-complicated logic on colorbar labelling.
The very complicated rules over colorbar labelling could lead to bizarre results in many cases, for example when a log plot was made with a small range or when the double-log range was used. The matplotlib default formatters are actually pretty intelligent and other cases should be handled manually by the user.