Allow reading of arbitrary dimension tipsy aux arrays (e.g. 1D or 3D)
This addresses issue #379.

For binary arrays, the number of dimensions are inferred robustly from the
file size, dtype, and number of particles.

For ascii arrays, the number of dimensions are estimated by estimating the
average size (in bytes) of a single data point and dividing the file size
by this and the number of particles.  This should be fairly robust except in
some edge cases.  A warning is raised if the estimate appears bad.