Allow reading of arbitrary dimension tipsy aux arrays (e.g. 1D or 3D)
This addresses issue #379.

There are several different file types, all which have dimensions inferred
using different methods:

 * For binary arrays, the number of dimensions are inferred robustly from the
file size, dtype, and number of particles.
 * For gzipped binary arrays, the uncompressed filesize cannot be directly
read for files >4GB, but in general it is possible to correctly estimate the
number of dimensions
 * For ascii arrays, the number of dimensions are estimated by estimating the
average size (in bytes) of a single data point and dividing the file size
by this and the number of particles.  This should be fairly robust except in
some edge cases.  A warning is raised if the estimate appears bad.
 * Reading of gzipped ASCII arrays has also been implemented, but the number
of dimensions cannot be estimated