Delay import of pynbody.plot
Until now, pynbody has auto-imported pynbody.plot when the main package is imported.

Importing pynbody.plot imports pylab which in turn is quite slow and
can cause problems for terminal-only applications. However, since pynbody
always auto-imported pynbody.plot, it would seem to be too destructive to
stop this behaviour. So this hack is the compromise and should be completely
transparent to most users.