Prevent Ramses particle loading from failing silently
Ramses particle loading could silently fail due to creating an empty array
that looked like a successful load to the framework.

Implemented immediate-term fix; in the long term, the framework should not
treat IOError in a load array request as expected/recoverable; a different
exception should be used to indicate an expected failure.