Fix bug with GadgetHDF masses
When the mass was stored in a GadgetHDF header, the wrong units could be
inferred in some circumstances.

This was due to the file units system being used to guess the right units.
The file units system in turn had been wrongly inferred, which in turn was
because multiple HDF array names ending in "Mass" were found, some of
which had different units.