Fix bug where sometimes family arrays would return the entire simulation
SimArray.sim is supposed to point either to the entire simulation, or if it is a family array,
it should point to the family to which it corresponds. In some circumstances, it was returning
the entire simulation despite being a family array which can lead to huge confusion around
indices.