Refactor the way that SubSnap works
This is quite a major change to the guts of pynbody, but does pass the current tests. The main purpose,
other than making things a little clearer, is to enable a new type of 'auto-copy' subsnap, which
copies a subset of the data rather than pointing to the live underlying data. This functionality in turn
will be used by tangos to enable a shared memory pynbody server.