Finish reverting shared memory support to posix_ipc module
The Python shared memory support is too overbearing and restrictive (this may be partially fixed in
Python 3.13). For now it makes sense to revert to using posix_ipc, while minimising the coupling to
the posix_ipc module in the hope of moving away from its use in the long term.

See #762 for discussion, and #752 for the attempt to implement pynbody's needs with Python's own
in-built shared memory support.