Fix issue #771 and associated discovered problems
The root cause of #771 is the strange way in which sink particles are handled in Ramses. They are stored in a
separate csv file with its own naming conventions.

The test data file stored masses as `msink` while the test presented in #771 stores the mass as `mass`. In the
pynbody test data file, one would end up with zeros for the mass of the black holes, whereas for the test data
in #771 one could easily end up with cryptic messages about the family array. To fix the pynbody test case,
name mapping has been enabled for the csv fields. To fix #771, the special-case handling of `mass` arrays in
ramses snapshots has been fixed.

To discover this I had to delve back into the SimSnap function _get_array_with_lazy_actions, untouched for many
years. I have documented this a bit better and removed what seemed like an erroneous claim to be 'calculating'
an array when actually loading it. This had a tertiary knock-on effect of highlighting a bug and an inefficiency
in the way that changa/gasoline starlogs are handled.