[WIP] improvements to KDTree
Currently non-working set of improvements to KDTree

Investigation has revealed that

1) the Wendland kernel tests were actually not testing anything at all
since they were loading density information from disk (!) rather than
deriving it

2) In reality, when working with a Wendland kernel float32 precision
 can be highly problematic because the kernel function is so steep.
 Noise at the 10^-8 level with the cubic spline kernel suddenly rises
 to the 10^-3 level with Wendland.

This meant that implementing optional double-precision KDTrees becomes
more important, but that in turn would be a waste of time if not coupled
to the longer-term work to modernize the KDTree (#770).

So, I have started work on that modernization but at the moment it is not
returning correct smoothing, for reasons that are unknown but may relate to
the initialization of the priority queue, or perhaps the re-initialization
of the priority queue when switching from one particle to another.