Fix KDTree/smoothing for doubles (float64)
This implementation is still broken for float32 arrays because in some places
SmoothingContext<double> is hard-coded for convenience. This now needs to be relaxed
and then it should start to work for floats as well.

Also at the moment, the code is around 40% slower than the old code, but note that the old
code was hard-coded to float32, whereas the new one is hard-coded to float64! So this may
not be a fair comparison, and there are also opportunities for optimizing the new code e.g.
by using different containers to underly the heap. (Profiling would seem wise first)