Major updates to SPH rendering pipeline
These are logical updates which do not affect the output (much).

Create a new renderers submodule of the SPH subpackage that expresses the logic
of rendering with pynbody, including projection, denoising, approximations,
parallelism and more.

Move such logic out of the plotting package.

While doing this work, discovered that approximate_fast is no longer faster than
direct rendering for a lot of test cases. Retained the functionality, but used
fewer levels. Probably this is to do with the rise of core numbers, and other
changes in modern architectures, but I didn't dig very much deeper.