Add (failing) acceptance tests for persistent transformations.
There is currently an inconsistency in how transformations apply to freshly loaded data.

For example, if one rotates a simulation but it has not yet loaded the velocity array, then one loads the velocity array,
it will have the wrong orientation.

Worse, if/when the rotation is then reverted, the velocity array will be "unrotated" even though it was never rotated
in the first place, leading to inconsistent data even after the transformation is out of scope.

These acceptance tests illustrate how this ought to be resolved (but since there is no code to support it they will
currently fail.)