Fix issue with inconsistent dtypes between derived/loaded arrays
This arose in the context of the images tutorial with gadgethdf, but
was potentially a general problem where if one derived a family array
then tried to load the same-named array into another family with a different
dtype, one would get an exception.

Confusingly for the user, if the order of operations were reversed (load,
then derive) things would be fine.

The solution is to relax how strict pynbody is about creating arrays
of different dtypes at family level.