Probable fix for long-standing issue on CI Windows tests
This may be down to the low resolution of the Windows time,
which could have led to name collisions in shared memory.

On dev machine, this wasn't an issue but on CI it was, perhaps due
to different time resolutions on different Windows versions.

Making the problem very hard to diagnose, name collisions were not
being detected by the Windows backend.

Hopefully it's now all fixed. Test added.